Create 'create_book.rb' file
