+++ data/labels.json
@@ -22,5 +22,13 @@
     "items": [
       474
     ]
+  },
+  {
+    "id": 623,
+    "name": "Sports",
+    "color": "Gold",
+    "items": [
+      247
+    ]
   }
 ]
